Better Tracks customer order count and include customer status
